requirement.txt
# Core ML Libraries
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
datasets==2.14.0

# AWS & Cloud
boto3==1.28.85
sagemaker==2.189.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10

# NLP & Text Processing
spacy==3.7.2
python-dotenv==1.0.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# ML Utils
scikit-learn==1.3.2
tqdm==4.66.1

# Web Application
streamlit==1.28.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
jupyter==1.0.0
jupyterlab==4.0.9
black==23.11.0
flake8==6.1.0

# Optional - for enhanced OCR
# easyocr==1.7.1  # Uncomment if needed